refactor: make currencySelectDropdown as console component
