+++ src/common/modules/dropdown/currency-select-dropdown/CurrencySelectDropdown.vue
