+++ src/services/cost-explorer/cost-analysis/modules/CostAnalysisQueryFilter.vue
@@ -70,11 +70,12 @@ import dayjs from 'dayjs';
 
 import { i18n } from '@/translations';
 
+import CurrencySelectDropdown from '@/common/modules/dropdown/currency-select-dropdown/CurrencySelectDropdown.vue';
+
 import CostAnalysisPeriodSelectDropdown
     from '@/services/cost-explorer/cost-analysis/modules/CostAnalysisPeriodSelectDropdown.vue';
 import { GRANULARITY } from '@/services/cost-explorer/lib/config';
 import { getInitialDates } from '@/services/cost-explorer/lib/helper';
-import CurrencySelectDropdown from '@/services/cost-explorer/modules/CurrencySelectDropdown.vue';
 import { costExplorerStore } from '@/services/cost-explorer/store';
 import type { Period } from '@/services/cost-explorer/type';
