+++ src/services/dashboards/widgets/WidgetsPreviewPage.vue
@@ -84,7 +84,8 @@ import { store } from '@/store';
 
 import { CURRENCY } from '@/store/modules/display/config';
 
-import CurrencySelectDropdown from '@/services/cost-explorer/modules/CurrencySelectDropdown.vue';
+import CurrencySelectDropdown from '@/common/modules/dropdown/currency-select-dropdown/CurrencySelectDropdown.vue';
+
 import DateRangeSelector from '@/services/dashboards/widgets/_components/DateRangeSelector.vue';
 import type { DateRange } from '@/services/dashboards/widgets/config';
 import { WIDGET_THEMES } from '@/services/dashboards/widgets/view-config';
